rename diffuser
